Move PyParams to bindings_ie
